requirements/development.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=./requirements/development.txt ./requirements/development.in
#
asgiref==3.3.4
    # via django
backcall==0.2.0
    # via ipython
click==8.0.1
    # via pip-tools
decorator==5.0.9
    # via ipython
django==3.2.3
    # via
    #   -r ./requirements/base.in
    #   django-filter
    #   djangorestframework
django-filter==21.1
    # via -r ./requirements/base.in
djangorestframework==3.12.4
    # via -r ./requirements/base.in
flake8==3.9.2
    # via
    #   -r ./requirements/development.in
    #   flake8-docstrings
flake8-docstrings==1.6.0
    # via -r ./requirements/development.in
invoke==1.6.0
    # via -r ./requirements/development.in
ipdb==0.13.7
    # via -r ./requirements/development.in
ipython==7.23.1
    # via
    #   -r ./requirements/development.in
    #   ipdb
ipython-genutils==0.2.0
    # via traitlets
isort==5.9.2
    # via -r ./requirements/development.in
jedi==0.18.0
    # via ipython
matplotlib-inline==0.1.2
    # via ipython
mccabe==0.6.1
    # via flake8
parso==0.8.2
    # via jedi
pep517==0.11.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.2.0
    # via -r ./requirements/development.in
prompt-toolkit==3.0.18
    # via ipython
psycopg2-binary==2.9.2
    # via -r ./requirements/base.in
ptyprocess==0.7.0
    # via pexpect
pycodestyle==2.7.0
    # via flake8
pydocstyle==6.1.1
    # via flake8-docstrings
pyflakes==2.3.1
    # via flake8
pygments==2.9.0
    # via ipython
pytz==2021.1
    # via django
snowballstemmer==2.1.0
    # via pydocstyle
sqlparse==0.4.1
    # via django
toml==0.10.2
    # via ipdb
tomli==1.0.4
    # via pep517
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.36.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
